requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pypdf
sentence-transformers
faiss-cpu
numpy
langchain
